[BF-DOCS-1] convert docs/layout.tsx to Server Component (#1000)
'use client' layout causes RSC flight transition stuck (expiredLanes) in Next.js App Router.
Extract all hooks/state/handlers to docs-client-layout.tsx ('use client'),
leaving layout.tsx as a pure Server Component wrapper.
